requirements.txt
requests
openai
python-dotenv
flask
flask_restful
flask-cors
